Refactor: dto 어노테이션 추가
