+++ src/main/java/org/nhnacademy/book2onandonbookservice/dto/common/BookContributorDto.java
@@ -1,10 +1,14 @@
 package org.nhnacademy.book2onandonbookservice.dto.common;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
+@AllArgsConstructor
+@Builder
 public class BookContributorDto {
     private Long id;    // book_contributor_id
     private String roleType;    // role_type
